Add CRUD and seeder for booking status enum

+++ Application/Interfaces/EnumMasterData/IEnumMasterDataRepository.cs
@@ -2,7 +2,7 @@
 
 namespace Application.Interfaces.EnumMasterData
 {
-    public interface IEnumMasterDataRepository : IRepositoryAsync<Domain.Entities.EnumMasterData.EnumMasterData, long>
+    public interface IEnumMasterDataRepository : IRepositoryAsync<Domain.Entities.EnumMasterData.EnumMasterData, int>
     {
     }
 }
